ASSEMBLER REPORTS PROJECT
This file, ASSEMBLER REPORTS.txt, is a job log from an IBM z/OS mainframe system. It was produced by running a JCL job on a Marist College z/OS system. The log contains information about the job's execution, including the user ID associated with the job, the steps executed, the resources used, and the job's return code.